Allow organisations to be archived
This adds a "Delete this organisation" link to the organisation settings
page. The link shows a confirmation banner and makes a request to API to
archive the organisation if the user clicks confirm.

This is similar to what we have for services, but a little simpler. Only
platform admin users can see the settings page, so no one else can
archive an organisation. We also don't create an event.